refactor(shared): update imports to use shared/config
