+++ src/components/organisms/reviewsWithInfiniteScroll/index.tsx
@@ -6,7 +6,7 @@ import { useCallback, useEffect, useState } from "react";
 import { getReviews } from "@/effect/reviews/getReviews";
 import { FilterBar } from "@/components/molecules/filterBar";
 import { Filters } from "@/entity/filters";
-import { reviewsSortOptions } from "@/constants/sortOptions";
+import { reviewsSortOptions } from "@/shared/config";
 import { getFormattedDate } from "@/shared/lib";
 
 interface ReviewDetailProps {
